requirements.txt
annotated-types==0.7.0
anyio==4.9.0
boto3==1.39.14
botocore==1.39.14
certifi==2025.7.14
click==8.2.1
distro==1.9.0
fastapi==0.116.1
h11==0.16.0
httpcore==1.0.9
httpx==0.26.0
idna==3.10
Jinja2==3.1.6
jmespath==1.0.1
MarkupSafe==3.0.2
openai==1.30.1
pydantic==2.11.7
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
s3transfer==0.13.1
six==1.17.0
sniffio==1.3.1
starlette==0.47.2
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.35.0
